LayoutTests/inspector/profiler/canvas2d/canvas-stack-trace.html

<html>
<head>
    <script src="../../../http/tests/inspector/inspector-test.js"></script>
    <script src="../canvas-profiler-test.js"></script>
<script>

var canvas;
var context;

function createCanvasContext()
{
    canvas = document.getElementById("canvas");
    context = canvas.getContext("2d");
    console.assert(context, "Failed to create a canvas context");
}

function doSomeCanvasCalls()
{
    context.beginPath();
    context.rect(0, 0, 100, 100);
    context.fillStyle = 'red';
    context.fill();
}

function test()
{
    var traceLogId;
    InspectorTest.enableCanvasAgent(step1);
    function step1()
    {
        InspectorTest.evaluateInConsole("createCanvasContext()", step2);
    }
    function step2()
    {
        CanvasAgent.startCapturing(didStartCapturing);
    }
    function didStartCapturing(error, id)
    {
        InspectorTest.assertTrue(!error && !!id, "Unexpected error: " + error);
        traceLogId = id;
        InspectorTest.evaluateInConsole("doSomeCanvasCalls()", didSomeCanvasCalls);
    }
    function didSomeCanvasCalls()
    {
        CanvasAgent.getTraceLog(traceLogId, 0, undefined, didReceiveTraceLog);
    }
    function didReceiveTraceLog(error, traceLog)
    {
        InspectorTest.assertTrue(!error && !!traceLog, "Unexpected error: " + error);
        InspectorTest.addResult("");
        InspectorTest.dumpTraceLog(traceLog);
        InspectorTest.completeTest();
    }
}

</script>
</head>
<body onload="runTest()">
<p>
Tests stack traces in the Canvas TraceLog results.
</p>
<a href="https://bugs.webkit.org/show_bug.cgi?id=107805">Bug 107805</a>
<canvas id="canvas"></canvas>
</body>
</html>
