LayoutTests/inspector/profiler/canvas2d/canvas-stack-trace-expected.txt

Tests stack traces in the Canvas TraceLog results.

Bug 107805 

{TraceLog}  alive:true  startOffset:0  #calls:4
  {Call}  functionName:"beginPath"  arguments:[]  sourceURL:"canvas-stack-trace.html"  lineNumber:19  columnNumber:13
  {Call}  functionName:"rect"  arguments:[0,0,100,100]  sourceURL:"canvas-stack-trace.html"  lineNumber:20  columnNumber:13
  {Call}  property:"fillStyle"  value:red  sourceURL:"canvas-stack-trace.html"  lineNumber:21  columnNumber:23
  {Call}  functionName:"fill"  arguments:[]  isDrawingCall:true  sourceURL:"canvas-stack-trace.html"  lineNumber:22  columnNumber:13
